third_party/WebKit/LayoutTests/fast/loader/recursive-before-unload-crash-expected.txt

CONSOLE ERROR: Blocked alert('onbeforeunload called, and iframe hasn't been added yet.') during beforeunload.
ALERT: Adding iframe
